+++ addOns/wappalyzer/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Updated with upstream Wappalyzer icon and pattern changes.
 - Updated the pattern parser to deal with Confidence or Version fields extending DOM patterns (for the time being they're ignored).
+- Updated the passive scan rule to be thread safe.
 
 ## [21.9.0] - 2022-02-03
 ### Changed